model/requirements.txt
flask==3.0.0
flask-cors==4.0.0
waitress==2.1.2
objsize==0.6.1
fleep==1.0.1
vosk==0.3.42
soundfile==0.10.3.post1
pydub==0.25.1
transformers==4.40.1
torch==2.6.0
gTTS==2.5.1
pyttsx3==2.90
moviepy==1.0.3
openai-whisper==20231117
requests==2.31.0